+++ internal/applications/filesharer/internal/utils/utils_test.go
@@ -60,7 +60,7 @@ func newTsHLSClient(pMsgSize uint64) *tsHLSClient {
 func (p *tsHLSClient) GetIndex(context.Context) (string, error) { return "", nil }
 func (p *tsHLSClient) GetSettings(context.Context) (hls_config.IConfigSettings, error) {
 	return &hls_config.SConfigSettings{
-		FLimitMessageSizeBytes: p.fMsgSize,
+		FPayloadSizeBytes: p.fMsgSize,
 	}, nil
 }