+++ internal/service/pkg/config/config_test.go
@@ -5,10 +5,10 @@ import "testing"
 func TestConfig(t *testing.T) {
 	v := uint64(10)
 	c := &SConfigSettings{
-		FLimitMessageSizeBytes: v,
+		FPayloadSizeBytes: v,
 	}
-	if c.GetLimitMessageSizeBytes() != v {
-		t.Error("limit message size bytes != v")
+	if c.GetPayloadSizeBytes() != v {
+		t.Error("payload size bytes != v")
 		return
 	}
 }