+++ internal/service/pkg/config/types.go
@@ -4,14 +4,14 @@ import "github.com/number571/hidden-lake/internal/service/pkg/app/config"
 
 type IConfigSettings interface {
 	config.IConfigSettings
-	GetLimitMessageSizeBytes() uint64
+	GetPayloadSizeBytes() uint64
 }
 
 type SConfigSettings struct {
 	config.SConfigSettings
-	FLimitMessageSizeBytes uint64 `json:"limit_message_size_bytes"`
+	FPayloadSizeBytes uint64 `json:"payload_size_bytes"`
 }
 
-func (p *SConfigSettings) GetLimitMessageSizeBytes() uint64 {
-	return p.FLimitMessageSizeBytes
+func (p *SConfigSettings) GetPayloadSizeBytes() uint64 {
+	return p.FPayloadSizeBytes
 }